refactor: credential validator for simple service ops
This commit refactors the credential validator to predominantly remove
mongo state based calls from the facade. While here the oppurtunity was
taken to integrate credential service into the facade so invalidation
and model credential status questions can be asked in one shot.

What this allows for is the facade to operate from the model content
that it has and not have to perform a lot of back and forth building
information up. It also allows the facade to become very simple with
business logic.