feat: model credential status
In the credential validator agent facade we used to have logic that
would piece together the credential key being use by the model and also
if the credential for the model was considered valid.

This stiching of information andl also the buisness logic working out if
the credential was set on the model resulted in a lot of logic that
could be moved down to the facade.

This commit moves all of this information down to the credential service
with well defined contracts the facade can rely on. This also allows
callers to stay scoped firmly on the facade and not wonder off side ways
in the controller.