refactor: expose credential validity with model
On the model view this exposes the column that indicates if a model's
credential is valid. The reason for exposing this information this way
is it keeps with the theme of getting callers to use the model view over
raw queries.

This stops potential problems from running into models that are not
activated. This was added to support the credential service providing
status information about a model's credential.