feat: add state and service methods to set controller api address
This patch adds the state and service layer methods to insert or update
the api address for a given controller ID. This method is going to be
used as part of the apiaddresssetter worker to set the controller's api
address.
Some types were fixed for machine.UUID as a side change.