feat: remove the pubsub from agentconfigupdater
The agentconfigupdater used pubsub to be notified if the controller
config was updated. This isn't required as we can just use a watcher
and not rely on pubsub to perform update guarantees.

We still have to migrate the StateServingInfo from the API caller,
then this would be another worker that is just using the dqlite
directly.