refactor: model infor create users
This commit refactors all the callers of model info create func to use
the new function signature where they don't have to supply the
controller uuid anymore.

This helps decrease the deps required on the callers. While here I also
removed the model agaent argument from model import args as it no longer
lives in the controller database.

I also renamed a case of model id back to UUID so that is inline with
the rest of the controller domain model work.