feat: add api address setter worker
This patch is the central change regarding the api address setter
worker, which is intended to be a replacement for the peergrouper
worker.

This commit is the actual worker, which spawns watchers for the
controller nodes, the controller config and the controller node's
addresses changes, so it can update the client and agent api addresses.