chore: handle end of k8s session
It was found that when ending the k8s session, closing the slave side of the pseudo terminal doesn't close the master and the user needs to press enter again to stop the session. To work around this, we send our own new-line as a hacky workaround.
It might be the case that we don't need to create our own pty.