chore: address flakey sshsession worker tests
Calls to the RemoveEphemeralKey key method are not always consistent because the function is deferred until after the client's connection is closed and the test has no way of synchronising with that event, so we allow the mock to be called AnyTimes().
Other flaky behaviour emerges due to not checking if the worker is alive so mock calls to `ControllerSSHPort()` are not satisfied until we check the worker is alive.