refactor(charms): cut over CheckCharmPlacement to DQLite
This endpoint used Mongo to iterate over the units on a application,
checking their machines.

Most of this facade should be encapsulated into the machine domain.
Create a stub method for this, and wire this up to the facade, and in
doing so, decouple the facade from much of Mongo application code.

This will break some limited aspects of refresh until this service
method is implemented.