feat(application): wire up Public and Private unit address functions
Replace unit.PublicAddress and unit.PrivateAddress in the uniter with
their domain equivilents. Add tests for the wired up functions. This
includes fully testing EnterScope.

Note that the NetworkInfo facade method is not at all complete. It
currently only gets the public address of the unit. Other information is
not yet available. A full wire up will be done as part of the Link Layer
Devices epic.