chore: change ordering all facades
Part of this work removed the credential manager facade and then later
add it back in via another commit. When adding the facade back in the
order was change and that still caused a delta for the change that we
don't want. This commit sets the order back to what was before.