fix: cloud region not being set to null
I noticed a bug in the model database for cloud region. When there is no
cloud region in use by a model the value must be NULL in the database.
This is how we model it in the controller database and how we should
also model it in the model database.

However when seeding information from the controller database to the
model database we represent the not set value as the zero value of a
string. This is the issue that has been fixed and it should now be set
to NULL.