refactor: model readonly owner information
When originally making the read only model table in the model database
we were only setting the credential owner on the model information and
conveying to users that this was the same as the owner of the model.

This is incorrect because credential owners can be different to that of
the model owner. It is also possible for model's to not have a
credential set.

On top of this we were only setting the user name and not the unique
uuid value of the owner. As we transition more over to Dqlite Juju
wishes to use more foundational id's to represent things like users
instead of natural keys like names.

We now store both bits of information to help the transition to uuids.