refactor: credential owner in model database
This commit fixes an issue in the model read only table where if no
credential is set for the model then the credential owner is still set
to an empty string.

What we needed was a way to set this value to null in the database how
it is supposed to be represented.