chore: fix flaky test TestSSHSessionWorkerMultipleConnections
The test was not correctly using mocks. The test creates 2 pairs of connections, one between the worker and the controller and another between the worker and sshd. When running the test with `stress` it could happen that the "wires get crossed" and the first pair of connections was associated with the second.
Now we synchronise these operations to ensure the test is always consistent.