feat(application): wire up unit upgrader facade with application
Wire up the unit upgrade facade with the application service. Get the
units machine ID from the application service rather than from the
domain.

The tests in this facade are not currently implemented. I will save
implementing these for the machine epic (unless anyone has an issue with
that) since any tests written now will still have to be moved
from using the machine state to the machine domain.