chore: drop unused param from state methods
The state methods AddIAASUnit and AddCAASUnit took in a charmUUID param
which is not used. These methods instead retrieve the charm UUID using
the application reference.

I suspect this is a hangover from when the domain made use of RunAtomic,
since we could then retrieve the charm UUID in the service layer.