feat: get user uuid for user name
This commit adds a new function to the user service for getting a user
uuid for a given user name. This as apposed to the current method where
all the information about a user is brought our when you just want the
uuid.

This work was done to support the model manager facade when listing
model summaries for users.