refactor: list model summaries in model manager
This commit introduces missing information and full implementation of
list model summaries on top of dqlite for the model manager.

This has been mostly a refactor of how we get this information. We know
that the information needed for making a model summary has to come from
both the controller database and the model database.

Specifically the information needed from the controller database is that
of the users access and information to build the model's status. Instead
of forcing the model manager to have to understand how to stich these
two sources of information together it is better if the model manager
compiles a list of uuid's it cares about then asks of each model info
service for the information as a one shot call.

As part of this commit I have also removed the list model tests to be
added back in at a later date.