refactor: model controller service
This commit refactors the controller model service to remove the two
methods that it had around obtaining model summaries for the entire
controller and also on behalf of a user.

Also added to the service is the ability to list the model uuids a user
access to. The reason for this change is the ability to generate model
summaries for a model are being shifted to the model info service. Now
all we need is to know what model uuids are required.