refactor: refactor of model state for summary
This commit refactors the model state implementation to remove the
original implementation that was done for model summaries and add in a
new per model implementation.

This new implementation meets what the service layer requires and allows
for all model summary information to be extracted.

While here the oppurtunity was taken to refactor for loops in
transactions.