refactor: apply feedback to model status work
This commit sets a bad owner tag on model status as it leaves the
facade. This is to make the facade work with the client until model
owner is refactored in the controller.

While here I also change the model type from a string to the core type.
Several updates around documentation removing the word basic as it
implies something that may or may not exist in the implementation.