refactor: the LifeSuspendedStatusWatchers
Combine the common code of the principalLifeSuspendedStatusWatcher and
subordinateLifeSuspendedStatusWatcher code into a
lifeSuspendedStatusWatcher. This reduces the potential problems if one
watcher is updated and not the other. It also makes clear where the two
watchers overlap.

Each unique watcher implements the GetMapper method, where the
difference in watching principal and subordinate applications can be
found.