feat: add draining methods to objectstore
This adds the draining methods for the objectstore flag. This
includes the addition of getting the current active draining info,
along with updating it. We only allow one active draining info
to be active at once, so there is no need to expose some of the
information outside of the service.