chore: fix documentation mistakes and misses
This commit applies PR feedback recieved to address some of the
documentation on struct types and the explanations that were given for
unit tests and what their purpose is for.

All this should be chore related and has no affect on functional
aspects.

This work is in relation to the list model summary support for the model
manager facade.