feat: remove pubsub worker
The pubsub worker forwarded messages to other controllers. We don't
need to do that anymore, as all left over publishing messages are
local only.

This is a massive step towards the removal of the pubsub library.

We want to use the database as the single source of truth and only
expand on to these sorts of eventual consistency methods when they're
required.