chore: disable unit upgrade precheck
Remove check that compares the unit CharmURL to the application
CharmURL to check if the unit is mid-upgrade. The CharmURl is no longer
written in mongo so this check always fails.

We disable this check because the upgrade functionality is going to be
rewritten in 4.0, so a new check will have to be wired up as part of
that future work anyway. This has been recorded in the upgrade epic.