refactor: epehemeral provider tracker runner
The provider tracker has the concept of an ephemeral provider that can
be constructed off of the state of the model at a given point in time
with no mechanisms for performing updates when model information
changes.

The reason this exists is to support provider resource creation when
establishing the model. This gets us out of the situation where the
provider tracker needs model information but that information gets
created in the same process that sets up the information needed for the
provider.