fix: check for context cancels on lease error
If WaitUntilExpired in WithLeader returns an error, check that the context
has not already been cancelled. Without this check, the random order of
select statements meant that either the WaitUntilExpire or the context
error could be returned (though the first error had a lower probability
of being returned as two selects had to go in its favour).

This patch makes the bahviour consistant by first checking the context
before returning the WaitUntilExpired error.