refactor(application): cut over endpoint bindings on client application
Cut over a few things to be backed with DQLite state in the client
application facade.

In particular, cut over our calls to EndpointBindings to a new service
method.

This required a small change to the state method getEndpointBindings.
In the DDL for application_endpoint and application_extra_endpoint, a
NULL value in space_uuid represents the endpoint is bound to the
application's default space. We now will in this default value.

This _could_ cause trouble to modelmigration, however, we have a check
in the import code to ensure we don't fill in the space_uuid if it's the
same as the application's default space. I have added more some tests for
this case.