test: fix flaky worker test
The TestConnectWithSameAddress worker sent a message down a channel in
the apiConnectHandler func, this was listened to in the test to check
that the apiConnectHandler had been called, when it was it then
proceeded to do another test. The problem was this second test relied on
the addresses it passed in having been written to an internal field in
the worker. This was a change that happened after and in a different
thread to the apiConnectHandler caller. So there was a possibilty the
addresses would not have been written by the time the second test read
them.

To fix this, instead of waiting in the test on a channel from used in
the apiConnectHandler, use the workers internal state. This is updated
after all the necessery changes have taken place, so we know it is safe
to wait.