feat: add getter for api addresses
This patch adds the service and state layer methods to retrieve
controller node API addresses.

Also, a new index has been added to ip_address on net_node_uuid, the net
nodes watcher in the app domain has been renamed to WatchUnitAddresses
to avoid exposing the net node concept when we only care about the unit
addresses.

As a side quest as well, the controller node API address insertion has
been improved so we only update the addresses if there is indeed a
change in them.