feat: delete unit
Adds delete unit method to the removal domain. We can remove the
application delete unit code once the removal code is wired up.

This copies a lot of the removal code from application, but cleans
some of the naming duplications.

Whilst doing this, we also delete the annotations, which required
adding a FK to the annotation unit table.