feat(application): add & wire up a watcher for deployer/firewaller
This commit primarily creates a new watcher to the application domain,
and wires it up to the deployer and firewaller workers.

This watcher is designed to shadow the old Mongo-backed WatchUnits
watcher, which emits the names of units whenever a unit is added or
removed from a specific machine. Continue emitting the names for
compatibility.

Since we're just interested in certain events, a set of auto-generated
triggers will not be sufficient for us. Instead, create some custom
triggers for a custom namespace. Reserve the first 10000 namespace ids
to be extra safe for future additional namespaces.

Wire this watcher up to the deployer facade and the firewaller directly.
This allows us to remove the common unit watcher facade.

Alongside this, make a number of small changes/fixes:
- Ensure that a principal unit exists before adding a subordinate
- Ensure that the unit_principal table is cleaned up when a subordinate
  unit is removed. This fixed a bug meaning we could not remove subordinates