feat: implements model name validator
This commit adds a missing model name validator to the core model
package. The regular expression rules around what forms a valid model
name have come from the juju/names package.

Includes a suite to test a set of invalid and valid model names to
try and identify changes to the regex over time.

This new validator will be used in the various model domains to support
creating and importing models.