feat: new modelmanager domain with service
This commit introduces a new modelmanager domain with an initial
implementation of the service for the domain. This domain exists to
provide controller level actions for managing the models within a
controller. Also apart of this service is the ability to create new
models within controller and import existing models into the controller.

Up until now this functionality has been located within the model domain
where it has sat along side model information fetching methods. Because
of the seperate concerns it makes sense to break out the management
aspects of controller models into a seperate domain.

While moving code from the model domain the opportunity has been taken
to also fix up documentation and make the naming of methods consistent
across the service and domain.

The design for this new service is one where it cannot directly be used
to interrogate a models information.