fix: use pebble v1.19.1
This brings in pebble v1.19.1 which includes a fix for how Notices and Changes
are pruned. In v1.19.0 changes might be pruned, but the associated notice would
then refer to a missing change. We increase the number of changes kept to 1000
(up from 500), and then ensure that notices are kept consistent with changes.

See https://github.com/canonical/pebble/pull/613 for more details.