feat: introduce strong types for cloud types
This commit introduces strong types for representing the type of a
cloud. Currently when talking about a cloud's type inside that of a
domain we use the string values and have no way to refer to a cloud type
by a stronger primitive.

In this commit the cloud domain now has a CloudType value that is based
off of the integer index in the DDL. The core/cloud package now
also includes a set of CloudType enums that are based off of the string
representation of the cloud type.

Tests have been added to the cloud domain to check first that the integer
enums in the domain are aligned with the values that exist within the
controller DDL. Another test in the cloud domain also checks that the
core cloud type string values align with the type string representation
of a cloud type in the controller DDL.

On top of this tests have been added to each provider to make sure that
the provider type value each provider uses to register itself in the
environ registry is the same as the value defined in core/cloud. This
now means all the way from the the DDL to the provider we have unit
tests to detect skew and drift of values.

The tests for the provider were done from the providers aspect of the
problem. If the test was to be conducted from any other part of the
code base it would bring in the provider package in as a dependency. This
would introduce the problem where we are unable to compile out certain
providers and their dependencies from the controller.

The work here is done in advance to support a new ModelManager domain
that is being developed. The ModelManager domain needs to be able to
look at the cloud type of a model and calculate a corresponding type
for the model. By now having this logic rely on the integer cloud type
enums it means less lookup and matching needs to be performed within
the database.

This pattern also makes the cloud domain and cloud type more
inline with the idiomatic approach Juju is going in with DQlite where
we prefer unique identifiers over that of strings.