chore: allow controller to restart without jimm reachable
JAAS provides 2 public keys, one for signing JWTs and another public key used when minting macaroons with 3rd part caveats.
If JAAS is not reachable, the Juju controller should remain operational. This was prohibited by the fact that we were fetching the public keys on startup in a blocking manner.
This change ensures the keys are only fetched on-demand and cached to keep things operational when JAAS is down.