feat(uuidgen): add UUID generation and testing utilities
This patch is to address the fact that we create a lot of UUID for a lot of things.
It allows to one-line the creation of utility function for strong typed UUIDs for usage in code and
in tests.

Generation provides, for each UUID types generated:
- method NewUUID,
- method ParseUUID,
- implementation for Stringer interface
- Validate method which return "errors.NotValid" if not valid

Also, in a subpackage "testing", it provides method for generating uuid in tests.