feat: add double writing of API addresses at bootstrap
This patch adds the insert of API addresses in the bootstrap worker.

To achieve this, a few changes were needed in the internal/bootstrap
package, namely around the deployer methods for them to return the
addresses needed in `CompleteProcess()` to be inserted.
Also, since we have the addresses at this point, we return them to the
bootstrap worker and they can be used as bootstrap addresses. This is
useful to remove a hardcoded `localhost` address that the bootstrap
address finder returned in the case of k8s (i.e. if we don't get any
address from the bootstrap address finder, then we use the one returned
from the bootstrap deployer): we guarantee having a correct address
always.

We need to continue writing to mongodb until we replace all the calls to
`APIHostPortsForClients` and `APIHostPortsForAgents` with the controller
node domain, and also until the apiaddresssetter worker is wired up.