fix: api addresses filtering by management space
While working in the bootstrap wiring of k8s addresses, I realised that
the logic for inserting API addresses in the controller node domain was
wrong/incomplete.
In the original implementation (currently still present in the bootstrap
worker for the legacy state), we always kept the addresses for the
agents if the management space is nil and if all the addresses were
filtered down, this way we ensure that agents can always reach the API
server.

This commit addresses these issues by updating the service layer method
`SetAPIAddresses` in the controller node domain, also changing the
signature to receive a pointer to space info (thus the changes in the
call sites.

Some tests were added to ensure this behavior.