feat: prevent data race in state pool closing
If the state pool is closing and we're attempting to iterate over
the pool, whilst it's still being written to, causes issues.

The solution is simple, create a shallow clone of the pool whilst
holding the lock and loop over that.

This is not a behavioural change, just the fact that we're not
iterating over the pool whilst it's still being written to.