fix: slice access without guard causes panic
The showformatter command attempts to be clever and break word
wrap a long line to a fix column width. Except that the calculation
for some words just isn't correct and it requires it to overflow
on to the next line. Except that the number of lines is calculated
upfront. The solution (**hack**) is to add one more line for when
this happens, so it can place the last word.

The real solution is to write a better algorithm to work out the
correct spacing for words with the space being removed from the
fields. Yet, this isn't used that often, and isn't in a hot path,
so I'm fine with shoving an additional line in at the end to solve
the problem.