refactor: remove core cloud type
Previously in the set of commits around introducing enums to represent
the various cloud types within a Juju controller I had defined string
based cloud type enums in the core cloud package.

We are actively trying to move away from types in the core package and
instead use the domain package for these values. Given this information
it makes sense to just have one CloudType type in the domain cloud
package.

Because this enum value is an integer and not a string a new String()
method needed to be introduced to convert the type to a string value
from an integer.

At this stage there is no reason why we need to convert from a string
type to that of an integer. I have left this for other implementers when
the time comes.

Also as part of this PR two int values were created to represent low and
high range of cloud type integers. These values can then be used to
validate if the integers are valid and also cover cases where the
database has had a new value introduced but no enum exists as of yet.