feat: validation support for agent stream
Adds support to the agent stream type in the model agent domain for
validating the enum without the need for state. This allows for the
service layer to perform pre-checks quicker and avoid a transaction for
simple validation.

This was done to support model upgrading where for some model upgrades
the caller wants to also set a new agent stream to use during the
upgrade.

This commit includes sub tests to make sure that the database values are
always considered valid. This adds to the test coverage for testing
drift between the database and that of the domain.