feat: support for upgrading model agent version
This commit introduces support to the model agent domain for upgrading
the target agent version of a model. What this allows for is for a model
to
be upgraded over time to meet new released versions of Juju.

Previously this logic was located within the modelupgrader facade and
handled both upgrading a single model and also that of the controller
model. The code paths for both of these operations are very distinct but
the implementation tried to make them follow similar patterns.

This new implementation only supports upgrade a model that isn't a
controller model and leaves the controller model upgrade implementation
to a seperate domain for a later date.

There is now 4 new funcs for upgrading. The first two functions allow a
caller to just upgrade the model to the latest version available with
the option to also change the agrent stream.

The second two functions allow a caller to upgrade the model to a
specific version while also changing the agent stream.

The set of validation problems that need to be addressed have all been
covered and include their own error types for which there exists special
handling cases today in the facade.

To help with testing I also introduced a new testing suite to just
perform upgrade based testing. This provides a better conceptual sandbox
for identifying the set of problems being checked in tests.

As part of the upgrade checks we need to confirm if agent binaries exist
for the new version. As this binary checking capability is still being
worked on the binary finder always returns true for the moment. A future
task will come along and wire this up correctly. Tests have been
included to validate that binary checking does work when it is
eventually wired up to the service.