feat(endpoint): cut over MergeBindings to DQLite
Wire up the MergeBindings client facade to write bindings to Mongo
and DQLite. We can cut over entirely, because we no longer write
bindings to Mongo all the time, meaning any Mongo reads are already
broken.

NOTE: The old MergeBindings method would verify when re-binding an
application that the machines that application is deployed to are
deployed to the target spaces. This check has not been implemented yet.

This required some refactoring to the application endpoint state layer.

When we create an application, one of the (many) steps is to insert
initial endpoint bindings for the application. This was performed by
insertApplicationEndpoints.

This method is quite complex, and would be very similar to a method to
update an application's bindings. So I have refactored the method
slightly to upsert the bindings, instead of just inserting them.

As a flyby, refactor some charm state methods. Many of these were
written before applications had a not null foriegn key constraint to the
charm table, so handled many impossible cases. Also, wire up an existing
private method. This results in massive simplifications.