feat: wire the api address setter worker
This patch wires the api address setter worker into the jujud-controller
machine manifolds, as a common manifold between IAAS and CAAS.

During QA a problem was discovered with the worker, which couldn't start
because of a wrong type returned from `GetDomainServices`. This was
fixed by separating the controller and controller model domain services
into separate injecting methods.