feat: dual write bootstrap api addresses
This patch adds the insert of API addresses in the bootstrap worker.

Regarding the changes to the address finder, as a continuation of the
work we have been doing to remove any "if caas", I have separated
both implementations. One difference in the case of K8s is that instead
of returning the hard-coded "localhost" address, the actual service
addresses are returned from the k8s broker. If the returned addresses
are empty then "localhost" is returned, for backwards compat.

We need to continue writing to mongodb until we replace all the calls to
APIHostPortsForClients and APIHostPortsForAgents with the controller
node domain, and also until the apiaddresssetter worker is wired up.

As a side cleanup: remove unused method GetCloudServiceAddresses() from
the application service layer. We use unit addresses (public|private)
instead.