chore: change UpdateCloudService to take provider addresses
The method UpdateCloudService used to take space addresses as input, but
this is wrong and we never use the space in the addresses.
This patch changes the signature to take provider addresses instead.

Taking provider addresses means that we no longer need to assign the
alpha space to the retrieved addresses from the providers during
bootstrap.