refactor: model agent machine util testing
This is a small commit to fix up the naming of the two util methods that
exist for model agent state around adding a machine to the model.

Several tests need to add one or more machines to the current model to
assert behaviour about the model agent state. These funcs were
previously called createMachine*.

It is more idiomatic with Juju testing standards to call them
addMachine* as the operation is to augment the current model with more
value and context.