fix: missing namespace for agent_version change
There was a custom trigger written for the agent_version table to
register model agent version changes in the change log. However the
trigger registration code failed to register the agent version namespace
for which the change events are assigned to.

This resulted in sql updates to the agent_version table failing with a
foreign key error.