feat: machines not on bases state
This commit defines the implementation for model agent state to
calculate the number of machines in the model not running on a defined set
of bases.

This state layer implementation is required for performing validation
steps during a model upgrade.

Included in this commit is a set of tests to validate the contract on offer.

I also took the chance to remove the plural s from the end of the
function signature. Reflection suggested that this was not needed and
sounds better as a function name without the s.