refactor(firewaller): drop Mongo-backed facade
The firewaller worker api used to call the controller facade
GetAssignedMachine to get the machine id a unit is assigned to. This
facade was backed by Mongo.

However, the firewaller have direct access to the application service.
So I have dropped the facade entirely, and instead use the service to
get the assigned machine directly.

As a flyby, restructure the firewaller to used machine names instead of
machine tags, since tags are an api thing, and we are trying to
gradually bypass the api for the firewaller.