feat: add method for retrieving the controller model UUID
This patch adds a service layer method for retrieving the controller
model UUID in the model domain. The state layer method already exists so
this simply exposes it as a service method.

As a flyby, this method is recruited in the api address setter worker,
and some minor changes (comments and moving of interfaces) are done on
the worker as well.