feat: provide state layer for setting model agent version
This commit introduces the state layer required for the model agent
domain to set (update) the model's target agent version.

This is required when updating a model to a newer version and is the
final step in the process after all of the upgrade checks have been
performed.

I took the time in this PR to also do a slight name change to one of the
database types used for getting a model's currently set target agent
version. The name change makes the type name inline with the other types
of the file.

Included in this commit is a suite of tests covering off on the known
edge cases of the state layer to assert the happy path cases.