feat: agent stream not valid error
To support upgrades of a controller it is common for the agent stream to
also be changed as part of the upgrade. We did not have a specific error
in the model agent domain to represent an invalid agent stream before
this.

This commit adds a new error to the model agent domain for indicating an
invalid agent stream back to the caller. This was done because
coreerrors.NotValid is required to signal another user input issue and
we are not able to overload the error type.