feat: resuse endcoding from machine
When performing a machine placement during deployment, we need
to set the statuses for both machine and instance as pending. This
is a pain, as we're using enum integers in the database. So we
need to map them accordingly. Rather than having this in two palces
and could potentially become out of sync, just use the same function.