feat(caasfirewaller): remove caasfirewaller controller facade
The controller facades are an odd thing. We have a set of api facades on
the controller which are only used by workers living on the controller
itself.

As such, these workers have direct access to the domains, so don't need
to go via the facades (which creates expensive network connections).

Refactor the caasfirewaller to perform all it's domain operations
directly on the domain services, rather than going via facades. This
means we can remove a lot of boilerplate api and facade code.

This was done to remove the caasfirewaller's dependence on Mongo. This
turned out to be the easiest (and best) way to do this.