feat: expose machine names when deploying
When deploying an application for IAAS, we assign a unit to a machine
if required. We need to pass that information back to the service so
it can record the correct historical event. The status history is only
available at the service level, hence all the mechanical changes.

This really cements the fact that we split the create application for
IAAS and CAAS. CAAS applications have no concept of machines and
applying this change doesn't touch CAAS methods in anyway.