fix: remove last vestiges of bash shebang path
This commit removes the last remaining shebang lines in bash scripts
that are assuming the path to the bash interpreter. This assumption
tends to break on POSIX systems that install bash in a location that
isn't the one hardcoded by the script.

The more portable and safer solution given the number of environments
Juju scripts have to run on is to use env to find bash on behalf of the
script.