chore: bump Pebble version to v1.19.2
This includes a fix for a panic that happened when Pebble was
restarted after a stop-checks operation.

This bug probably won't affect Pebble in a Juju K8s context, as when
Pebble exits, K8s will restart the container (the state file is
effectively ephemeral), but it seems good to update anyway.