feat: domain service implementation for upgrading controllers
This commit introduces the service layer for upgrading controllers in a
cluster to a new patch version when available. Previously to this commit
the logic around performing upgrades has lived in the modelupgrader
facade supporting both upgrades of the model and also supporting
upgrades to the controller.

We now consider these operations distinct with separate implementations.
Provided in this new service is the ability to upgrade a controller to
the latest patch version available or to a version picked by the
controller. Each of these variants also a stream implementation that
supports changing the agent stream for the controller as part of the
operation.

Checks are performed on all controller nodes in the cluster to confirm
they are in a position to be further upgraded. In the facade
implementation this service is replacing all models in the controller
were also checked before the upgrade.

This included checking for version, supported bases and also LXD
version. Because we do not support minor or major upgrades of the
controller anymore it doesn't make any sense to re-implement these tests
under this domain. As patch versions must not drop support for what was
supported before.

There exists the possibility of a half write error between the
controller and model databases. Comments and special test cases have
been constructed to show that this is a recoverable problem for the
caller. It also informs any future changers of this code what the
contract expectation is.

Not implemented in this commit is either of the controller or model
states. Also there exists the need for a specialist binary finder to
support the operations that has also not been done.